energy form API refacto
