--- src/handlers/energy.ts
@@ -1,52 +1,0 @@
-import { Request, Response } from "express";
-import { Database } from "../types/database.types";
-import { createClient } from "@supabase/supabase-js";
-
-const supabaseUrl = process.env.SUPABASE_URL;
-const supabaseKey = process.env.SUPABASE_ANON_KEY;
-
-if (!supabaseUrl || !supabaseKey) {
-  throw new Error("Missing Supabase environment variables");
-}
-const supabase = createClient(supabaseUrl, supabaseKey);
-
-type Energy = Database['public']['Tables']['energy']['Row'];
-
-
-export const saveEnergyResult = async (req: Request, res: Response) => {
-  const { projectId } = req.params;
-  const body : Energy = req.body;
-
-  try {
-    const { data, error } = await supabase
-      .from("energy")
-      .insert({
-        id : body.id,
-        izi_response: body.izi_response,
-        id_project: parseInt(projectId),
-      });
-
-    if (error) throw error;
-
-    res.status(201).json(data);
-  } catch (error) {
-    res.status(500).json({ error: error });
-  }
-};
-
-export const getEnergyResult = async (req: Request, res: Response) => {
-  const { projectId } = req.params;
-
-  try {
-    const { data, error } = await supabase
-      .from("energy")
-      .select("*")
-      .eq("id_project", parseInt(projectId));
-
-    if (error) throw error;
-
-    res.status(200).json(data);
-  } catch (error) {
-    res.status(500).json({ error: error });
-  }
-};
