+++ src/handlers/energy/index.ts
@@ -1,0 +1,13 @@
+import { createClient } from "@supabase/supabase-js";
+
+const supabaseUrl = process.env.SUPABASE_URL;
+const supabaseKey = process.env.SUPABASE_ANON_KEY;
+
+if (!supabaseUrl || !supabaseKey) {
+  throw new Error("Missing Supabase environment variables");
+}
+export const supabase = createClient(supabaseUrl, supabaseKey);
+
+export * from "./choice";
+export * from "./results";
+export * from "./form";
