+++ src/routes/energy/choice.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { saveEnergyChoice } from "../../handlers/energy";
+
+const router = Router();
+
+router.post("/", saveEnergyChoice);
+
+export default router;
