+++ src/routes/index.ts
@@ -2,18 +2,15 @@ import { Router } from "express";
 import projectRoutes from "./project";
 import scanRoutes from "./scan";
 import { getCredential } from "../handlers/log";
-import energyFormRoutes from "./energy-form";
-import energyMixRoutes from "./energy-mix";
+import energyRoutes from "./energy/index";
 
 const router = Router();
 
 // /api/uuid
 router.get("/uuid", getCredential);
 
-
 router.use("/projects", projectRoutes);
 router.use("/scans", scanRoutes);
-router.use("/energy-form", energyFormRoutes);
-router.use("/energy-mix", energyMixRoutes);
+router.use("/energy", energyRoutes);
 
 export default router;
